Add staging hash to URL
